+++ webapp/src/components/responder-portal/submssion-form-content.tsx
@@ -7,7 +7,7 @@ export default function SubmissionFormContent() {
     const { data } = useSubmissionContext();
     if (!data) return null;
     return (
-        <div className="mt-12 w-full">
+        <div className="w-full">
             <FormRenderer form={data.form} response={data.response} isDisabled />
         </div>
     );